refactor: Improve error handling in type parser
